+++ packages/expo-audio/ios/AudioModule.swift
@@ -19,7 +19,7 @@ public class AudioModule: Module {
       try setAudioMode(mode: mode)
     }
 
-    AsyncFunction("setIsAudioActiveAsync") { (isActive: Bool)  in
+    AsyncFunction("setIsAudioActiveAsync") { (isActive: Bool) in
       try setIsAudioActive(isActive)
     }
 
@@ -45,9 +45,14 @@ public class AudioModule: Module {
 
     // swiftlint:disable:next closure_body_length
     Class(AudioPlayer.self) {
-      Constructor { (source: AudioSource?, updateInterval: Double) -> AudioPlayer in
+      Constructor { (source: AudioSource?, updateInterval: Double, enableLockScreenControls: Bool, metadata: [String: Any]?) -> AudioPlayer in
         let avPlayer = AudioUtils.createAVPlayer(source: source)
-        let player = AudioPlayer(avPlayer, interval: updateInterval)
+        let player = AudioPlayer(
+          avPlayer,
+          interval: updateInterval,
+          enableLockScreenControls: enableLockScreenControls,
+          metadata: metadata
+        )
         AudioComponentRegistry.shared.add(player)
         return player
       }
@@ -153,19 +158,20 @@ public class AudioModule: Module {
       }
 
       AsyncFunction("seekTo") { (player: AudioPlayer, seconds: Double) in
-        await player.ref.currentItem?.seek(
-          to: CMTime(
-            seconds: seconds / 1000,
-            preferredTimescale: CMTimeScale(NSEC_PER_SEC)
-          )
-        )
+        player.seekTo(seconds)
+      }
+
+      Function("updateMetadata") { (player: AudioPlayer, metadata: [String: Any]) in
+        player.updateMetadata(metadata)
       }
     }
 
     // swiftlint:disable:next closure_body_length
     Class(AudioRecorder.self) {
       Constructor { (options: RecordingOptions) -> AudioRecorder in
-        guard let cachesDir = appContext?.fileSystem?.cachesDirectory, let directory = URL(string: cachesDir) else {
+        guard let cachesDir = appContext?.fileSystem?.cachesDirectory,
+          let directory = URL(string: cachesDir)
+        else {
           throw Exceptions.AppContextLost()
         }
         let avRecorder = AudioUtils.createRecorder(directory: directory, with: options)
@@ -249,7 +255,8 @@ public class AudioModule: Module {
     }
 
     do {
-      try AVAudioSession.sharedInstance().setActive(isActive, options: [.notifyOthersOnDeactivation])
+      try AVAudioSession.sharedInstance().setActive(
+        isActive, options: [.notifyOthersOnDeactivation])
       sessionIsActive = isActive
     } catch {
       throw AudioStateException(error.localizedDescription)