+++ packages/expo-audio/ios/NowPlayingInfoController/NowPlayingInfoKeyValue.swift
@@ -1,0 +1,13 @@
+//
+//  NowPlayingInfoKeyValue.swift
+//  SwiftAudio
+//
+//  Created by Jørgen Henrichsen on 28/02/2019.
+//
+
+import Foundation
+
+public protocol NowPlayingInfoKeyValue {
+    func getKey() -> String
+    func getValue() -> Any?
+}